LayoutTests/inspector/console/console-format-style-expected.txt

CONSOLE MESSAGE: line 8: %cBlue!.
CONSOLE MESSAGE: line 9: %cBlue! %cRed!
Tests that console logging dumps properly styled messages.

Blue!. console-format-style.html:8
Styled text #0: color: blue;
Blue! Red! console-format-style.html:9
Styled text #0: color: blue;
Styled text #1: color: red;
